+++ changelog.md
@@ -5,6 +5,7 @@
 - **Fixed:**
   - Tray context menu hide show/hide all windows if only 1 window exists for simple menu
   - Fixed error when theme file not found
+  - Fixed default paths for themes, sounds and logs for none window systems
 - **Changed:**
   - Update better-sqlite3 8.1.0 to 8.2.0